+++ temporalio/nexus/_operation_context.py
@@ -711,6 +711,7 @@ async def _start_nexus_operation_workflow_update(  # pyright: ignore[reportUnuse
     update: str | Callable,
     arg: Any = temporalio.common._arg_unset,
     args: Sequence[Any] = [],
+    wait_for_stage: temporalio.client.WorkflowUpdateStage,
     update_id: str | None = None,
     result_type: type | None = None,
     rpc_metadata: Mapping[str, str | bytes] = {},
@@ -718,6 +719,8 @@ async def _start_nexus_operation_workflow_update(  # pyright: ignore[reportUnuse
     run_id: str | None = None,
     first_execution_run_id: str | None = None,
 ) -> temporalio.client.WorkflowUpdateHandle[Any]:
+    if wait_for_stage != temporalio.client.WorkflowUpdateStage.ACCEPTED:
+        raise ValueError("Only ACCEPTED wait stage is supported")
     # Default update ID to the Nexus request ID for retry-safety (matches sdk-go).
     update_id = update_id or temporal_context.nexus_context.request_id
     workflow_handle = temporal_context.client.get_workflow_handle(
@@ -728,7 +731,7 @@ async def _start_nexus_operation_workflow_update(  # pyright: ignore[reportUnuse
             update,
             arg,
             args=args,
-            wait_for_stage=temporalio.client.WorkflowUpdateStage.ACCEPTED,  # hardcoded as nexus only supports async updates
+            wait_for_stage=wait_for_stage,
             id=update_id,
             result_type=result_type,
             rpc_metadata=rpc_metadata,